requirements.txt
asyncio
aiohttp
pyodbc 
pandas
pyarrow
python-dotenv
apache-airflow==2.8.0